moved sendtocuda function
